+++ apx_sdk/CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed MSRV to 1.80.0.
 - Migrated to `thiserror` 2.0.
 - Replaced `required_components` argument in `parse_http_signature_rfc9421` with a boolean flag.
+- Don't try to extract content digest if request method is not POST, PUT or PATCH.
+
+### Security
+
+- Return `NoDigest` error if digest header is not present in POST, PUT or PATCH request.
 
 ## [0.24.0] - 2026-05-13